[feature:client] Add script sources for create form
