+++ index.html
@@ -3,6 +3,7 @@
   <head>
     <meta charset="utf-8">
     <title>Pinwall</title>
+    <script src="//cdn.tinymce.com/4/tinymce.min.js"></script>
     <link rel="stylesheet" href="./client/assets/css/normalize.css" >
     <link href="https://fonts.googleapis.com/css?family=Josefin+Slab:400,600" rel="stylesheet">
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" integrity="sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7" crossorigin="anonymous">
