+++ package.json
@@ -27,6 +27,7 @@
     "react-dom": "^15.2.0",
     "react-redux": "^4.4.5",
     "react-router": "^2.5.2",
+    "react-tinymce": "^0.5.1",
     "redux": "^3.5.2",
     "redux-form": "^5.3.1",
     "redux-persist": "^3.2.2",
